[INTERNAL][FIX] Table: Edge cases in NoData handling and keyboard navigation
The table is rendered without rows if there is no data and the NoData text is
visible. If the NoData text is now hidden (showNoData set to false), the content
area of the table is empty. The table is invalidated in this case so that rows
are rendered.

If there are no rows, for example because the row count is set to 0, the tab
elements around the row container are removed from the tab chain.

Change-Id: If452af7d01fb7ce2d1f457c2765ed42a7ffa5a04