[FIX] sap.m.Select: Problem with changing selection is fixed
Problem: After chosing a selection from the drop-down list by pressing
"Enter", navigation with arrow keys between different selection is not
possible anymore. Also, when tabing on the Select and again pressing
"Enter", the Select cannot be opened with "Space" press.

Solution: It seems that, upon "Enter" press, the hidden select menu with
options is getting opened as this is the default browser action. This
leads to focus stuck, until "Enter" is pressed again. So now we prevent
the default browser action, when "Eneter" is pressed.

Change-Id: I124addd2f3c74c24d84e5753ef55cb9cb7edbf9b
BCP: 2070308334