[INTERNAL] sap.ui.integration.widgets.Card: Change the way extension is created
Problem:
It is possible to have the same card (same manifest) more than once.
If there is 'extension' in the manifest, the module defining that
extension will be loaded only once, because of the way how module system works.
This would result in such cards using 1 extension instance.

Solution:
In order to keep card-extension relation 1:1, the extension
should export a class, not instance.

Change-Id: Ifa6dde4de4a967f016fbf2d5e4443de6a5cd91a0
BCP: 2070307801