[INTERNAL] Demo Kit: Fixed Tools page image overlap
Problem: The image in the cards in the Tools page, overlap with the
card info in some browsers (Safari).

Root Cause: The image is sometimes automatically set to be displayed
as block, where as the info is flex.

Solution: We set the card info to be displayed as block. This way we
guarantee that on all browsers, the info will be rendered with priority
and the image won't overlap.

Change-Id: I2e1d977ba701ff7b9bd048af7c05e072eaea18e2
JIRA: BGSOFUIPIRIN-4771