[FIX] Field: update editMode and multipleLines only on new rendering
if edit mode is bound, it might switch multiple times in binding phase.
This should not trigger multiple changes of the internal control.

As the change of the editMode invalidates the Field, a new rendering
will be triggered.

as editMode and multipleLines are often changed together the
determination of the new inner control schould only happen once on new
rendering. (Otherwise an unneeded Control might be created and directly
destroyed.)

Change-Id: I138a3a1361a1ba4d97acbd601dc0e706230c6dcc
BCP: 2070284113
CR-Id: 002075125800003508972020