[INTERNAL] m.Table: disable selectAll for no data
Due to the change Iedee0f376fd59dc37b77ffbefccea75185f91432, we prevented
disabling of the private control that are created internally by the table.
This led to an issue that the selectAll checkbox did not disable when there
are no items in the table. This issue has now been fixed. We disable the
selectAll checkbox of the table when there are no visible items in the table.
BCP: 2080330611

Change-Id: I941ccd073b70c1a70bbd0bbb956e53e81045dbe3