[FIX] sap.f.DynamicPage: Endless snapping/unsnapping is fixed
Problem: The header of the DynamicPage is endlesslly snapping and
unsnapping.

Soluiton: When determing whether the header can be snapped with
scroll, we check if the maxScrollPosition is higher than the
snappingHeight. However, in some cases, when maxScrollPosition is 0,
the snappingHeight comes as a negative value, due to substracting the
bottom padding from the header/title height. As a result, we wrongly
consider that we can snap with scroll.
Now we make sure there are no negative values.

Change-Id: If4c83744deb862ab5115d13f73a8e0c4e64684d9
BCP: 2070328159