[INTERNAL][FIX] ListBase: Show NoData if no items after rebind with growing
If growing is enabled and there are no items after a rebind, the table was not
invalidated. As a result, the items before the rebind were still visible
(items were destroyed though). Invalidation is required in this scenario to
remove the DOM of the old items and render the NoData element.

Change-Id: Iab0340987c45f7c40295e7335ed77002c9326f4e
BCP: 2070329271