[FIX] P13nConditionPanel: Tooltip shouldn't exists
Issue: In conditional ValueHelpDialog the ComboBoxes
are having tooltip which leads to double reading of their
value by JAWS.

Solution: Add a check if the control to which we add
tooltip isn't the box container of the ComboBox control.
By guidelines we shouldn't use tooltip on the ComboBox
since "The content of the tooltip would be redundant.".

Change-Id: I845964d0b3651993e921d0e55157ea3eda96ade7
BCP: 2080274790