[INTERNAL] sap.uxap.ObjectPage: Section is invalidated when internal title is set
Background: A section always starts with internal title level set to TitleLevel.Auto
and if needed later a new value is assigned and set to the aria-level attribute in the renderer.

Problem: If no value is assigned to the internal title or if is assigned but the control does not
get invalidated aria-level attribute is equal to the "o" from the "Auto" string.

Solution: Invalidate the control after a new value is set to the _internlaTitleLevel property
and ensure that a value different than the possible aria-level values is returned from the
_getAriaLevel function.

Change-Id: I89b72bb1340ad7638f2004e25ee702a8b3bb65e9
BCP: 2070280316