[INTERNAL] DemoKit: Ensure datatables config is loaded
Problem: In some cases the dataTablesConfig file is still not
loaded when a topic is matched and tables are not pushed to the
aWaitingDataTables array.

Solution: Ensure tables are pushed to the array even if the config
is not loaded and then load the config file with plugin request sequence.

Change-Id: I7473748003f329ffa605fbfba62a8e3992072395
JIRA: BGSOFUIPIRIN-4706