[INTERNAL] DemoKit: Improve collapsible elements
Problem: A display="block" was set to the expanded collapsible element
which was introducing issue with table's default display property "table".

Solution: Reviewing the logic again, it was noticed that the display
property is not needed when element is expanded as the display="none" is
no longer applied.

Change-Id: I819febfa8f73f18c117196230c4e812a351f5f43
JIRA: BGSOFUIPIRIN-4680