[FIX] sap.f.FlexibleColumnLayout: Contextual settings is passed
Problem: When applyContextualSettings is called from
FlexibleColumnLayout (before navigation), ObjectPageLayout and
DynamicPage execute their own logic, without passing the
contextualSettings object further in the call chain. This breaks Table's
popin logic, as the correct contextualSettings are not propagated to it.

Solution: Now the contextualSettings object is passed and used in the
call chain.

Change-Id: Ib5836e69c797e545889d594ed14b28aa13c5a2ee
BCP: 2070331833