[FIX] sap.f.GridContainer: Focus is now correct via Drag and Drop
Problem1:
When item is dropped(keyboard or mouse) in an empty grid
itemNavigation is not working correctly,setting the focus,
beacause of ghoast element.

Problem2:
When grids with only one items are rendered,
focus goes to the last grid with one element.

Solution:
Force focus only when focusItem function is called.

Change-Id: Ieb6d25432bd7eb86fa3b2bea617accac271fafe9
BCP: 2070340624
BCP: 2070340054