[INTERNAL] sap.m.Table: update colspan for popin when column visibility changes
When a column is set to visible=false and the table contains popins, the
colspan attribute for the popin cell was not updated, and this lead to an
unexpected layout of the table, where the popins were wider than the table.
This is now fixed where we correctly update the colspan for the popins when
column visibility is changed.
BCP: 2070268600

Change-Id: I69c8c237c4ec7be210daae2c195759b4d46cd63b
CR-Id: 002075125800003793002020