[FIX] sap.m.Breadcrumbs: Corrected width measurement
Background:
The control measures the widths of its aggregated controls
(and the width of its own container) using $().outerWidth(true)
in order to calculate if there is enough room to render all
aggregated control outside the overflow

Problem:
The $().outerWidth(true) uses rounding, which sometimes leads
to a result when the sum of the aggregated controls appears to
exceed the container width by 1px. But the 1px difference comes
only as a result of rounding (because we sum the rounded values
of each child).

Solution:
- Use getBoundingClientRect to measure the width more precisely;
- Use a separate mechanism to obtain the element margins
  (to add any margins to getBoundingClientRect().width,
  as already implemented successfuly in sap.m.OverflowToolbar)

Change-Id: I2cde11de5c209eafee1597b6c94899e77d73225f
BCP: 2070292533