[FIX] sap.m.TimePicker: now change event is fired when the input field is cleared (oData v4/TimeOfDay)
* until now, when oData v4 is used and type used for TimePicker is TimeOfDay, change event was
not fired when the value of the TimePicker were cleared. This happened because of TimeOfDay behavior,
which parseValue Method returned null when the value was empty string. Later this provoked calling of
setValue() method of the TimePicker with value null (because it is different than current one ""),
which led to setting of _sLastChangeValue property to "", and in fact cleared the value previously set.
Then, when fireChangeEvent() method was called ,it didn't fired the change event because it assumed
that there was no change in the value (new value equals to last changed value).

* now this is fixed, and in setValue() method _sLastChangeValue changes to new value only if it is
different that the current value of the control.

Change-Id: I5aaa61eefefc81321c8c8839ee4159e9e6a32666
BCP: 2070354898