[INTERNAL] sap.ui.fl provides the parentVersion on saving draft changes
To instruct the back end to save draft changes, currently the
parentVersion is only a indicator to store these as a draft.
With this commit, another version than the current active can be the
basis for the draft changes.
Valid options for the parentVersion parameter:
* sap.ui.fl.Versions.Draft(=0): the already existing draft should be
extended
* sap.ui.fl.Versions.Original(=-1): the draft should have no ancestor
* any activated version: a new draft is created with the passed changes;
this will discard an existing draft
* any non valid number will result in an error

Change-Id: I7ac199002335a316736194258cadd2bc3bb764ed
JIRA: CLOUDIOI-21