[INTERNAL] sap.ui.rta.toolbar.Adaptation hide the restore in case the "Original App" can be reactivated
The restore of the Adaptation for Key Users is no longer done by the
"Reset" button, but by "reactivating the 'Original App'".
Therefore the button is hidden in these cases.
Currently the binding of the visibility is aimed to the featureToggle
and on the removal of the feature toggle, the binding must be switched
to the general versioningEnabled flag in the model.

Change-Id: Ie8dd928d6953cd1c3a917213c423e21baedac602
JIRA: CLOUDIOI-183