[FIX] sap.ui.model.analytics.AnalyticalBinding#_getKeyIndexMapping: use available service key index
Fix computation of service key index for an index of the displayed data.
If the service key is already known for a certain index, take that,
otherwise compute the new index.

http://localhost:8080/testsuite/resources/sap/ui/test/starter/Test.qunit.html?testsuite=test-resources/sap/ui/core/qunit/internal/testsuite.models.qunit&test=AnalyticalBinding

PS1: tests for existing code, POC for the fix and red test
PS2: small refactoring
PS3: fixed TDD
PS4-5: review comments

Change-Id: Idc9540868bc93f98ed1dc24c564ec68173f8c52a
BCP: 2070284104