[INTERNAL][FIX] sap.m.CheckBox: Control aligned with latest HC specs
By VD recommendation, the HC flavours for the Fiori 3 (Quartz) theme
should be copied from the Quartz Light flavour and then the deltas should
be applied as specified in the visual specification for the HC flavours.

This way we get the exact look as in the Quartz Light and Quartz Dark
flavours - in the HC flavours with a few adjustments made.

Additionally, the font-size CSS property of the checkmark icon is
increased from 10 to 12px in the base file, since this recommendation
applies for all themes.

Change-Id: Ie00a118ac7190a094fc6fc7d0c5d81c994e690ba
BCP: 2080370709