[INTERNAL] ListBase: swipe functionality fixed
The ListBase was added to the InstanceManager as a dialog. In this case,
The InstanceManager also expects the ListBase to implement the getCloseOnNavigation
function to be implemented. Since this was not done so far, when the swipe is
triggered and browser back navigation is done, this threw an error in the console.
To fix this issue the ListBase is now added as a Popover to the InstanceManager
which also fixes the issue when back navigation is performed.
BCP: 2070377198

Change-Id: I0ee0d4c2e6389a513e260ea57bdb19bcdbe4a8cb