[FIX] sap.m.GenericTag: Truncate issue is fixed
Problem: When a property of the ObjectNumber ("value" aggregation) is
updated, some of the GenericTags, placed in OverflowToolbar in
DynamicPage, are getting truncated.

Solution: Now GenericTag is subscribed for any changes of the
ObjectNumber, so that, whenever something is updated, it dispatches an
already implemented invalidation event to the OverflowToolbar. This way,
the OverflowToolbar invalidates itself and its size caches and
calculations again the space which is needed.

Change-Id: Iaff558abd990bc69b3f8274b49ad829dc87c6dd9
BCP: 2070384949