[INTERNAL][FIX] Table: The click event of the Select All checkbox not fired when text is selected
In order to allow selection of cell content, the table prevents firing the click event when some
text is selected. Because the text selection of the Select All checkbox was disabled, the previous
selection was not removed and the Select All click event was not fired. Therefore enabling text
selection on the Select All checkbox

Change-Id: I23db16cb65e10bb32ea9d46fe5c928fde1ba4fe9
BCP: 002075129400006537092020