[INTERNAL] Integration Cards: The case when the host never resolves a promise is handled
Introduced a new utility called Utils.timeoutPromise().
It will return a Promise which rejects if timeout is reached
before the original promise is resolved.

The utility is now used in ContextModel.js and in Destinations.js
where we wait for promises from the host.

Now even if a promise inside ContextModel is rejected or has timed out,
the card will continue initialization so that we can have better fault tolerance.

JIRA: BGSOFUIRODOPI-2446
Change-Id: I278d4488fdce3c1029972fd74a4ee86a130887cb