[FIX] sap.m.OverflowToolbar: Cutting of buttons is fixed
Problem: There are some predefined device-dependent CSS classes, which
depend on media queries and show/hide controls with CSS. In some cases,
upon device orientation change for example, some controls (previously
hidden) become visibile, but their width is already cached as 0, so that
wrong calculations are made.

Solution: The controls' sizes cache is reset whenever there is a device
change, based on media queries, so that calculations are made with the
new values.

Change-Id: I38c1f2e7aa5724a1effb1f46f141f91fc2d85e96
BCP: 2070351993