[INTERNAL] Card Explorer: Component card samples are fixed
Solving the following problem:
When a Component card sample from the Card Explorer is used in
SAP Workzone it shows errors for parameters not being translated.

The errors were caused due to the component was loading
the manifest second time and was parsing it again.
Solution: Setting "manifest: json" should not be used.

Now the samples are updated so that they don't use this setting.

Change-Id: Idc8a4006f43e7d5c3fd8f6b671199717a71487fa
BCP: 2070402435