[INTERNAL] sap.m.Input: DisplayListItem removed
In order to make all inputs consistent the DisplayListItem
is removed from the Input. Previously, this item
was used to render items with additional text.
Now, if an item has an additional text it is mapped
with the info property of the ListItem, like in the
ComboBox control.

Change-Id: I05905c67c99b68112dccf1ba53c7f22659bd6a63
JIRA: BGSOFUIRILA-2793