[FEATURE] sap.m.ProgressIndicator: Added information popover
Now, if displayValue is truncated, the cursor becomes a pointer (hand)
when the user hovers over the ProgressIndicator. The user is also given
the opportunity to click on it and see an information popover.

The information popover (sap.m.Popover) consists of text (displayValue)
and a close button.

Change-Id: If36094630804f314bc2b08e789cc94b91d0dc3e0
JIRA: BGSOFUIPIRIN-4672