[INTERNAL] sap.ui.core: remove unused dependencies
When dependencies are implicitly required due to side effects, move the
dependency to the end of the dependency list, remove the import
parameter from the AMD factory function and document the side effect
next to the dependency, e.g.
  "sap/ui/core/Core" // provides sap.ui.getCore()
or
  "sap/ui/core/mvc/Controller" // provides sap.ui.controller
or
  "sap/ui/core/util/ExportRow" // convenience dependency for legacy code

Exception: for unused renderer dependencies, specify them as renderer in
the class definition.

Change-Id: Idb0e07f5aeb9f27cd95452583209e89e3b4d5204