[FIX] sap.m.NotificationListGroup: Collapse button no longer loses focus after being pressed
When a NotificationListItem's OverflowToolbar Popover is opened, pressing on the
collapse group button no longer gets its focus changed to the
NotificationListItem's OverflowToolbar expand button.
This overrides the behavior of OverflowToolbar's focusing after its Popover gets closed.

When this control is used in a Popover, this behavior no longer causes the whole popover to get closed.

Change-Id: Ice1d4c88c99406d0b7eb7308a39f9c06a6fdd8fe
SNIPPIX: 33020
BCP: 2070386915