[INTERNAL] Support Assistant: Presets popover is no longer closed on Chrome
Issue: When the "reset" button is pressed, the ResponsivePopover automatically closes
on Chrome. This is a known issue, which should be handled by applications.
Now we handle it, so the popover doesn't close anymore.

Opa5 Adaptations:
Opa tests were considering only the Chrome case, now they are adapted.
As result, there are no longer failures on Firefox and IE11 browsers.

Change-Id: Ie40f25485b24f6d550b9015fe54a5018a92f506e
JIRA: BGSOFUIRODOPI-2408