[FIX] sap.ui.integration Editor: Lists reloaded on dependency change
If a list within the editor was requested via a destination it is
not refreshed if the destination changed.

Also fix below issues:
a. the additional settings missing issue in visualization in BaseField.js
b. the selected option showing issue in icon select.
c. the navigation issue in icon select.

Change-Id: Ic9d6daecde5819e9ec7908bdb7059308fb0e0cb1
CR-Id: 002075125800004607102020
BCP: 2070425097