[INTERNAL] Patcher: Replace DOM instead of remove and insert
 - Patcher: Node name change is now done via single replaceChild call
for the root node instead of removeNode and then insertBefore. This is
very common when the visibility of the control is changed.
 - Patcher: Object.keys call is not necessary to get remaining
attributes.
 - RenderManager: Invisible placeholder DOM node of the control should
be used by the Patcher to reuse this DOM without checking the visible
property of the control since the rendering reason can the visible
property itself. This will help to create DOM nodes via DOM APIs when
the control visibility is changed.
 - InvisibleRenderer: getDomRef API is added as a convenient way to get
the placeholder DOM of a control.

Change-Id: Ia5cc277eaabf630fe142ecec2786884a90c25150