[FIX] sap.m.MultiInput: Failing QUnit test on IE now fixed
The test "45. KeyboardEvents: onsapdelete when MultiInput has value"
was failing in IE as it was not able to find an instance of the Tokenizer
aggregation in the MultiInput at the time of the unit run.

The reason is that before each test of the module a jQuery .focus()
method was called on every MultiInput after instantiation. That method
manipulates the DOM with a delay and is sometimes executed after the
afterEach() call on the test.

This jQuery focus() call is now replaced with the framework's .focus()
method to prevent such timing issues.

BCP: 2070436601
Change-Id: I225d5b23cd946ac9d710d48060d506258a4fff7c