[INTERNAL][FIX] Remove version parameter on exit in case of backend-only Draft
When determining the reload method on exit we used to only check for
dirty draft changes. Since we can now switch versions while still having
a draft we need to check for not only dirty changes but also a backend
only draft - so that we remove(reload) on exit - so that when
re-entering RTA we dont have a version parameter in the URL which will
result in not loading the potential draft.

Change-Id: I8d45fb3ac650660536e2d01aec7e4e30bd4af20f