[FIX] DemoKit: Documentation deep link issue fix
Problem:
When we open the documentation section from a deep link that contains a hash
(for a subSection of the displayed docu page), the master tree failed to expand.
This happened because the corresponsing master-tree item could not be found by id,
as the id was still suffixed with the hash of the subSection

Solution:
Corrected the function that preprocesses the topic id to ensure it strips
any hash (to a subSection).

Change-Id: I93d7821a903e9330300226ebb0fb8b133282e379
BCP: 2070321937