[FIX] DemoKit: Prevent link URL corruption
Background:
The URLs of links are transformed at runtime
(from SEO to non-SEO format for non-SEO enabled versions).
Only the documentation links are skipped, as these are
currently already transformed at build-time.

Problem:
A link from the documentation (that should be skipped)
was not identified as such, because it contained nested
elements, so the check was made on the wrong element.

Solution:
Ensured the check on the type of the link is made on the
actual anchor element (rather than on its child elements).

Change-Id: I8941d5d28bbe093a7fa16c0e3f34e3fc5aef2b6e
BCP: 2070444719