[FEATURE] Demo Kit: Search suggestions
Search suggestions are added on Demo Kit`s custom search control.
The shown suggestions are according to the page the user is on, for
example in the Documentation section the suggestions are from the
Documentation, in the API Reference section the suggestions are from the
API Reference texts.

Also, there are options for navigating to each category in search page
results.

JIRA: BGSOFUIPIRIN-1559
Change-Id: I3ab15386cbef503f8d5b68f73e23f6329e687288