[INTERNAL] PropertyHelper: Provide access to property infos via facades
Instead of reading the original property info objects directly, the information
is accessed via facades which provide functions. It is still possible to access
the original info object with #getRawProperty for special use-cases.

With this, the property helper gains more control over the attributes and their
values. The intention is to reduce the risk of misinterpretation. This also
allows controls that implement their own PropertyHelper subclass to easier alter
the information, while maintaining a stable API that is used by shared modules.

Change-Id: I6de1d0ab0d09dbd407565a87daf56009dcddbc32