[FIX] sap/ui/util/isCrossOriginURL: only make relative URLs absolute
URI.js doesn't ignore URNs when using the second parameter which
indicates to make the first parameter absolute to the second. The
browsers' native URL API does. As we need to continue to use URI.js for
the IE11 support we need to ensure to make only relative URLs absolute
to document.baseURI.

Change-Id: Icf7ca598a25ad0ebe836b5337c399ebe8b5f3883
BCP: 2070458190