[INTERNAL] sap.uxap.ObjectPageLayout: Removed overstyles of ObjectNumber
By VD ObjectNumber should be used with the "sapMObjectNumberLarge"
CSS class applied when it is a part of the ObjectPageHeaderContent.
No over-styles are needed from the ObjectPageLayout side, thus they are
now deleted.

The "sapMObjectNumberLarge" CSS class is applied to all of our Demo Kit
samples where we have ObjectNumber within the ObjectPageHeaderContent.

Change-Id: Iae4059aeec53b269ff190c58407913e264abb87a
BCP: 2070361931