[INTERNAL][FIX] sap.m.OverflowToolbar: Fixed automatically closing of the popover
Problem:
When sap.m.ComboBox is in the overflow area and it is opened the overflow popover
is closed.

Root cause:
New "open" property is introduced in the sap.m.ComboBoxBase which was not mark
as no invalidation property.

Fix:
Now, "open" property is marked as no invalidation property.

BCP: 2070461267
Change-Id: I5ad5730f29d5824e60a11a3d01db298103c8b4d2