[FIX] sap.m.OverflowToolbar: Fix rounding issue with GenericTags
Problem: When there is an invalidation event fired from the GenericTag
control, OverflowToolbar fires a _contentSizeChanged event to the
DynamicPage, in order the flex-basis attribute to be set to "auto", so
that the updated GenericTag to occupy the needed space. However, this
way, the width of the OverflowToolbar is equal to the exact sum of the
width of all its content controls. In some cases, due to browser rounding
specifics, it appears that there is 1px which is not enough, so one
control goes in the Popover menu and an overflow button appears.

Solution: Fix the rounding issue by adding 1px content size tolerance,
when checking if all content of the OverflowToolbar fits in.

Change-Id: I434efb010f01d8a27bb15eb7aa3da95b33c6f8da
BCP: 2070141690
CR-Id: 002075125900004440022020
(cherry picked from commit 03a9da476dac89786095ab6a2aa793385d3c978b)