[INTERNAL] unified.calendar: fix wrong usages of 'new' operator / ctor
CalendarDate.fromLocalJSDate() returns a new object, not a constructor.
Calling new on the result is useless and fails when migrating
CalendarDate to an ES6 class.

Additionally, when a constructor is to be called as a function, not as a
constructor, then the call should not be a direct call (MyClass()), but
via 'call' or 'apply' (MyClass.call(this, ...)). And it should be
executed on the concrete class name, not via the generic property
'constructor'. This is the only pattern that can be easily migrated to
ES6 classes (where class constructors no longer can be called as
functions).

Change-Id: I0faa8c86667d3e72dbaa6f823758679dd7469805