[INTERNAL][FIX] Table: Get contexts at correct index range (variable row height)
If variable row heights are enabled, it can happen that the table triggers 2
requests during initialization. The Binding#getContexts call on refresh did not
consider the additional row that is added for variable row heights.

Fixed mode: The issue is partially fixed. If the "refresh" event of the binding
is fired during initialization, before the variable row height mode is enabled,
the issue still occurs (the additional row is still not considered). It is
currently only possible to enable variable row heights after initialization,
because the feature is still private and there is no property available.

Interactive mode: Is similar to the fixed mode, but is not considered in
this change, because this mode is not used in Gantt.

Auto mode: The issue does not occur, because the request length is
calculated based on the screen height.

Change-Id: If99882e6317467f5208939daaf98d7216afc95a7
BCP: 2070435649