[INTERNAL][FIX] Icon Explorer: Fixed visual misalignments
Problems:
1 - On very narrow window size, there is a white stripe between
the header and the tiles on the main page.
2 - When the window is resized vertically, the "SAP Business Suite" tile
appears taller than the rest of the tiles and ruins the whole look and
feel of the application.

Root Causes:
1 - In the narrow use case, the header becomes taller than the
background image.
2 - A CSS class is missing, which takes care of the height of the tile.

Solutions:
1 - We force the header to be with a maximum height of 444px, which
is the height of the background image.
2 - We add the "BlockLayoutCellHeight" CSS class to the "SAP Business
Suite" tile in the application logic in the Home view of the Icon
Explorer.

Change-Id: I7bea32043b334a90f9cceeca0faf88ffac112ef7
JIRA: BGSOFUIPIRIN-4907