[FIX] sap.f.DynamicPage: Prevent hiding the header on size chage
Background:
--By UX rule, when the header becomes bigger than a certain
threshold, it can no longer be snapped in the *non-scrollable*
title-area and should be automatically moved to the topmost
part of the *scrollable* content-area.
--Since moving the header to the topmost part of the *scrollable*
content-area affects the current scroll position, the user will
perceive a visual offset of the scrollable content => to avoid
this the page also scrolls back by the same amount of pixels
as the height of the newly added header.

Problem:
If the header was *expanded* before moving it to the content area,
then scroll adjustment described above causes the big header to become
hidden from view with scrolling => the header becomes undesirably snapped

Solution:
When the "headerExpanded" is "true", we prioritize the header state
(rather than the content position) => we skip the adjustment of
the scroll position to preserve the header state.

Change-Id: Ie5cd0467c256d1d7b37ebb65b4106013bfa0246b
BCP: 2070400160