[INTERNAL][FIX] sap.uxap.ObjectPageLayoutSection: Corrected handling of focus
Problem:
Based on configuration of the sap.uxap.ObjectPageLayout, when the section is focused
there is an extra stop in tab chain.

Root cause:
Internal rules when the tabindex should be set to "0" are based on the sections count.

Fix:
Rules that control the tabindexes of the subsections now are based only on the visible
subsections count.

BCP: 1980074248
Change-Id: I5f459f075a5b26dc98377a86a2e74e0e3d462c34