[INTERNAL] sap.m.TextArea: Visual tests adjusted
- Some of the samples are deleted, since they have shown
absolutely the same use case as the others available.
- Some of the tests that have not illustrated specific
use cases are also deleted.
- Now, there are samples and tests for all the value states,
showing the standard use case, the one with long value state text
and the one containing a link. For that samples, the whole screen is
captured in order to ensure that the value state text will be visible.
- A button is added in the Page's header in order to
be injected custom css to the body element that hides the
caret when the visual tests are run.
- All the instances are captured on focus.

JIRA: BGSOFUIRILA-2932

Change-Id: If59a438a19c54b872095c1633cc114207d90fbca