[FIX] core.mvc: views should be constructed via factory, not constructor
Adds a note to the documentation of the existing view implementations,
explaining that their constructor should not be used directly.
Application or framework code should rather use one of the corresponding
factories.

The constructors can't be hidden (@hideconstructor), as their
documentation is needed by the tool chain that produces the metadata for
XML code completion / XML validation.

Fixes: https://github.com/SAP/openui5/issues/2998
Change-Id: I5cd0e1c22feb4e7f6441f7a56ff9222f42ae02fd
BCP: 2070280315