[INTERNAL] sap/ui/fl/LayerUtils.getCurrentLayer no longer accepts a flag
The LayerUtils.getCurrentLayer had a flag to return "USER" whenever a boolean was passed with true.

This property was only required (and variable) in the sap/ui/fl/Change & Variant. At that point, the class handles this logic itself while every other consumer can just call the function or use sap/ui/fl/Layer.USER instead of passing a parameter.

Change-Id: I190bb5e51f46d5394802a0d13678f9c62eb33429