[FIX] sap.m.ValueStateHeader: Width calc based on the popup removed
When the popover gets resized while opening there is a moment in time
that it is with width 100%. In rare edge cases if controls that
calculate width based on the width of the Popup can catch it in that
moment. It is a timing issue, it also depends on the containers
of the ValueStateHeader.

Issue is reproducible only with maxSuggestionsWidth set to the Input
as it makes the popup to be resized onBeforeOpen leading to the timing
problem.

Now the ValueStateHeader does not dynamically take the width of the
Popup setting it as a fixed pixel width to itself. It has an auto
width, to always fit its parent container, and avoid cases like this one.

Unit tests are not available as it is very hard to reproduce it
in qUnit env.

BCP: 2070466087
Change-Id: Iac0e49fd07c8c459e27aa13757ddeb6043940c1f