[INTERNAL][FIX] sap.ui.rta Show correct popup when pressing on Discard
With the introduction of an additional discard dialog, the text message
property was named as another property. There was a commit(5025492)
which renamed the new property to make the two messages unique again -
but the wrong text key was replaced.
This commit will fix this which will result in showing the correct popup
again.

Change-Id: Id300ae63b619fd4ecfdaf3b48c1d1edeee02ae40
CR-Id: 002075125800005361322020
BCP: 2070450841