[INTERNAL] sap.m.MultiInput: Visual test adjusted
- A button is added in the Page's header in order to
be injected custom css to the body element that hides the
caret when the visual tests are run.
- All the controls are captured on focus when the visual tests are run.
- The test page is refactored. The unnecessary samples are removed -
the ones that did not show specific behaviour and the ones which
were not part of the page content. The sample is sctructured better, in order
to prevent page scrolling when the visual tests are run.
- Visual tests are added for the different use cases that were not covered.

JIRA: BGSOFUIRILA-2932
Change-Id: I524f78680fca6ec266825d4d48644befe8c5d527