[FIX] P13nConditionPanel: set the focus to the operation field
The problem:
With the new design the keyField is hidden but when a new
row is created it trys to set the focus on it.

The Solution:
Now when a new row is created check if the keyField is
visible and if not set the focus to the operation field.

BCP: 2070460316
Change-Id: I75abd26ea49ef2ea231c31288dd7f36b9dd42555