[INTERNAL] sap.ui.model.odata.ODataMessageParser: add generic message for unsupported status codes
If there is a status code less than 200, between 300 and 399, or greater
than 600, we do not know how to parse the response for messages, so we
add a generic error to the message model.
Some browsers return status code 0 if network connection is not
available.

PS1: Done
PS2: Removed tests and mock data for status code 900 introduced with
     change 973866
PS3: review comments, removed obsolete warning
PS4: review comments

Change-Id: I30af72b4fdbc0d6b5958057b73fe6281815f3322