[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Fixed overlapping of header visual indicators
Problem:
In some cases pin/unpin and snap/expand buttons overlap to each other.

Root cause:
Due to performance optimizations when the shotHeaderContent property doesn't
invalidate the sap.uxap.ObjectPageLayout and recalculations if the header visual indicators
should be shown are not done.

Fix:
Now, when the showHeaderContent value is changed recalculations are triggered
manually to be done.

BCP: 2070437846
Change-Id: I6fa9359f20da659ebc9766be1b35687cdca790b8