[INTERNAL] introduce sap/ui/fl/apply/_internal/flexObjects/Variant and enhance isReadOnly
The Variants require additional properties which are part of the sap/ui/fl/Change.
The Change.js is used to also create instances of variants.
This commit introduces its own Variant.js to separate this functionality.

In addition the isReadOnly provided for the consumer sap/ui/comp/smartvariants/SmartVariantManagement.js not a direct consumable result.
The leftovers are now also part of the Variant.js functions.

Change-Id: Id370fc507d5069a3452489203eadfb79bcb31d60
JIRA: CLOUDIOI-195