[INTERNAL]AdaptationMixin, AdaptationFilterBar: fix descendant cleanup
As the AdaptationFilterBar might be initialized prior to placing it in
a Dialog, setting it as dependent is not sufficient in terms of cleanup.

E.g. after it has been used in a p13n UI the parent might change.
The according 'removeItem' hook also appeared to be called, but the
actual 'continue' logic (implemented in ItemBaseFlex) currently needs
to be done manually in 'AdaptationFilterBar'.

BCP: 2080386012
Change-Id: Ie17a3bc317c477b25baaf383950adcfbebd9a2a2