[INTERNAL][FIX] sap.ui.fl - extension point registration for fragments
On registration of extension points in fragments that are created in
the controller of an application during runtime, sometimes the fragment
is not attached to the target view. As the target view is required for
the extension point registration we need to determine the target view
from the controller attached to the fragment.

Change-Id: I3af3a43ad3ffa18ab54ab22d99e67d11c3273956
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>
BCP: 2040790263
CR-Id: 002075125900004202362020