[FIX] sap.m.Select: Fixed behavior in InputListItem
Problem: When the Select is inside InputListItem and it's open, when we
click on it in order to close it, the focus is applied to the InputListItem
and the Select reopens itself.

Root Cause: There is a specific logic in the ItemNavigation, where the
clicked element (being the arrow or the label of the Select) or some of
its parents is supposed to be with jQuery ":sapFocusable" pseudo selector.
The label and the arrow as well as their parents up until the InputListItem
itself, aren't with the ":sapFocusable" pseudo selector. This causes
the focus to be applied to the InputListItem and the logic for
opening/closing the Select to fail.

Solution: One of the requirements for the pseudo selector to be applied, is
to have "tabindex" attribute on the given DomRef. This is why we apply
"tabindex" attribute with value "-1". By applying this attribute, we are
not including the room DomRef of the Select in the tab chain, but we
programatically comply with the pseudo selector requirements. Therefore
the root DomRef (parent of the label and the arrow) appears as focusable
element in the problematic logic in the ItemNavigation class.

Change-Id: Id1dbbe6d14481d563ea65a76890484a9c8b3b800
Fixes: https://github.com/SAP/openui5/issues/3077
BCP: 2070482964